moved rendering to a system. To ensure that the rendering system is being run in the main thread, I added the concept of "Sync systems", so that when a system group is created it's possible to specify whether it's possible to run it on a separate thread or not.

+++ src/renderer/Window.zig
@@ -74,7 +74,7 @@ pub fn destroy(self: Window) void {
     c.glfwTerminate();
 }
 
-pub fn getTime() f32 {
+pub fn getTime() f64 {
     return c.glfwGetTime();
 }
